refactor: moved e2e tests to dedicated test file
